+++ puppeteer/index.js
@@ -14,7 +14,6 @@ const path = require('path');
 
 
 // Observations:
-
 // Perks are part of the StorageClass class, from what I saw, they are not globally defined, they are instantiated and added to characters on demand.
 // You could create a "custom" perk with whatever value and the game would count it as valid, but to actually have an effect it
 // would need to be checked at some point with this.hasPerk, otherwise it just sits there.
@@ -384,6 +383,7 @@ const path = require('path');
                     StatusEffectShades: getGlobalsByPrefix('STATUS_'),
                     TailGenital: getGlobalsByPrefix('TAIL_GENITAL_'),
                     Upbringing: getGlobalsByPrefix('UPBRINGING_'),
+                    ValidColors: {},
                     ValidFlags: {},
                     ValidTypes: {}
                 };
@@ -414,6 +414,36 @@ const path = require('path');
                 for (let i = 0; i < pantyDataArr.length; i++) {
                     pantyData[pantyDataArr[i][0]] = pantyDataArr[i][1].panty;
                 }
+
+                // #endregion
+
+
+                // #region Colors
+
+                const globalSkinColors = window.GLOBAL.SKIN_COLORS;
+
+                var validSkinColors = [];
+
+                for (const skinColorType in globalSkinColors) {
+
+                    // 0: the actual color value
+                    // 1: some kind of pretty print description
+
+                    const colorArrays = window.GLOBAL.SKIN_COLORS[skinColorType];
+
+                    if (colorArrays.length) {
+                        colorArrays.forEach(colorArray => {
+                            validSkinColors.push(colorArray[0]);
+                        });
+                    }
+
+                }
+
+                validSkinColors = validSkinColors
+                    .filter((color, index, self) => self.findIndex(color2 => (color2 === color)) === index)
+                    .sort((l, r) => l.localeCompare(r));
+
+                global.ValidColors.Skin = validSkinColors;
 
                 // #endregion
 
@@ -645,10 +675,8 @@ const path = require('path');
     * @param {RegExp} regex
     */
     function getValidBodyTypesDefault(regex) {
-        return getValidBodyParts('BodyType', regex, (match, groupIndex) => {
-            return groupIndex == 2 && match.toLocaleLowerCase() !== 'kaithrit';
-            // For whatever reason, the game sets the 'TYPE_KAITHRIT' type for the [Kui-Tan Lt, XO Defender, XO Gunner] creatures
-            // even though 'GLOBAL.TYPE_KAITHRIT' doesn't exist
+        return getValidBodyParts('BodyType', regex, (_, groupIndex) => {
+            return groupIndex == 2;
         });
     }
 
@@ -657,11 +685,8 @@ const path = require('path');
     * @param {RegExp} regex
     */
     function getValidBodyTypesForArray(regex) {
-        return getValidBodyParts('BodyType', regex, (match, groupIndex, matches) => {
-            return ((!matches[4] && groupIndex == 2) || (matches[4] && groupIndex == 4)) && match.toLocaleLowerCase() !== 'kaithrit';
-            // 2 => simple check, 4 => array check
-            // For whatever reason, the game sets the 'TYPE_KAITHRIT' type for the [Kui-Tan Lt, XO Defender, XO Gunner] creatures
-            // even though 'GLOBAL.TYPE_KAITHRIT' doesn't exist
+        return getValidBodyParts('BodyType', regex, (_, groupIndex, matches) => {
+            return ((!matches[4] && groupIndex == 2) || (matches[4] && groupIndex == 4)); // 2 => simple check, 4 => array check
         });
     }
 
@@ -703,12 +728,6 @@ const path = require('path');
     // Array assignment|equality checks
     obj.ValidTypes.Penis = getValidBodyTypesForArray(getTypeCheckForArrayRegex('Cock', 'cocks', 'cType'));
     obj.ValidTypes.Vagina = getValidBodyTypesForArray(getTypeCheckForArrayRegex('Vagina', 'vaginas'));
-
-    // Other
-
-
-    // TODO:
-    // dNip - should be the same as pen but should check to be sure
 
 
     const validBodyTypesEnd = Date.now();
@@ -1020,6 +1039,73 @@ const path = require('path');
     // #endregion
 
 
+    // #region Colors
+
+    // haha, yet again another system that relies on string comparisons that aren't defined anywhere (except for skin)
+
+    /**
+    * Get valid colors, using regex and boolean callback
+    * @param {RegExp} regex
+    * @param {Function} callback
+    */
+    function getValidColors(regex, callback) {
+
+        var validColors = [];
+
+        util.execRegexOnContents(regex, (match, groupIndex, matches) => {
+            if (callback(match, groupIndex, matches)) {
+                validColors.push(match);
+            }
+        });
+
+        validColors = validColors
+            .filter((color, index, self) => self.findIndex(color2 => (color2 === color)) === index)
+            .sort((l, r) => l.localeCompare(r));
+
+        return validColors;
+
+    }
+
+    /**
+    * Get valid colors using the default implementation
+    * @param {string} name
+    * @param {string} type
+    */
+    function getValidColorsDefault(name, type) {
+        return getValidColors(getColorCheckRegex(name, type), (_, groupIndex, matches) => {
+            return (!matches[4] && groupIndex == 3) || (matches[4] && groupIndex == 4);
+        });
+    }
+
+    /**
+    * Get a regex that can be used to check for valid colors
+    * @param {string} name
+    * @param {string} type
+    */
+    function getColorCheckRegex(name, type = 'Color') {
+        return new RegExp(`(\\.${name}${type}(!|)={1,3}"([\\S ][^"),]+)"|"([\\S ][^"),]+)"(!|)={1,3}[\\w]+\\.${name}${type})`, 'g');
+    }
+
+
+    console.log('\ngetting colors');
+    const colorsStart = Date.now();
+
+
+    ['Hair', 'Lip', 'Eye', 'Fur', 'Scale', 'Nipple', 'Penis', 'Vagina'].forEach(name => {
+        obj.ValidColors[name] = getValidColorsDefault(name.toLocaleLowerCase());
+    });
+
+    obj.ValidColors.Penis = getValidColorsDefault('cock');
+
+    obj.ValidColors.SkinTone = getValidColorsDefault('skin', 'Tone');
+
+
+    const colorsEnd = Date.now();
+    util.printOperationTime('colors', colorsStart, colorsEnd);
+
+    // #endregion
+
+
     // #region Misc
 
 
@@ -1042,6 +1128,21 @@ const path = require('path');
     // #region Manual clean
 
 
+    // #region Colors
+
+    ['FIXME', 'NOT SET', 'PLACEHOLDER', 'no', 'none', 'unknown'].forEach(name => {
+        for (const colorType in obj.ValidColors) {
+            const colors = obj.ValidColors[colorType];
+            const index = colors.findIndex(c => c === name);
+            if (index > -1) {
+                colors.splice(index, 1);
+            }
+        }
+    });
+
+    // #endregion
+
+
     // #region Key Items
 
     [''].forEach(name => {
@@ -1060,6 +1161,23 @@ const path = require('path');
     // #region Counts
 
     const objLength = o => Object.keys(o).length;
+
+    const nestedObjLength = (root, key,) => {
+
+        const lengths = {};
+
+        Object.keys(root[key]).forEach(k => {
+            const items = root[key][k];
+
+            const length = items.length || objLength(items);
+
+            lengths[k] = length;
+        });
+
+        return lengths;
+
+    };
+
     const arrLength = a => a.length;
 
 
@@ -1068,8 +1186,6 @@ const path = require('path');
 
     // Global
     counts.Global = {};
-    counts.Global.ValidFlags = {};
-    counts.Global.ValidTypes = {};
 
     Object.keys(obj).forEach(k => {
         if (!k.toLocaleLowerCase().startsWith('valid')) {
@@ -1077,12 +1193,9 @@ const path = require('path');
         }
     });
 
-    Object.keys(obj.ValidFlags).forEach(k => {
-        counts.Global.ValidFlags[k] = objLength(obj.ValidFlags[k]);
-    });
-    Object.keys(obj.ValidTypes).forEach(k => {
-        counts.Global.ValidTypes[k] = objLength(obj.ValidTypes[k]);
-    });
+    counts.Global.ValidFlags = nestedObjLength(obj, 'ValidFlags');
+    counts.Global.ValidTypes = nestedObjLength(obj, 'ValidTypes');
+    counts.Global.ValidColors = nestedObjLength(obj, 'ValidColors');
 
 
     // Codex